Working on main game play system #3

+++ Assets/Resources/DataBaseConfigs/ShieldConfig/ShieldConfig.cs
@@ -25,7 +25,7 @@ namespace Sources.DataBaseSystem
     public class LevelInfo
     {
         public string Id;
-        public float Hp;
+        public int Hp;
     }
 
     [Serializable]
